fix(scoring): clear Sonar reliability bug + duplicate imports (QG green for #151)
- EvaluateurDashboardService: cast operand to long in
  plusMinutes(DUREE_STATION_MIN + GRACE_PERIOD_MIN) at both call sites
  (java:S2184 int-overflow). This was the lone reliability issue dragging
  the New-Code Reliability Rating to B and failing the Quality Gate.
- ExamenParticipationController: remove two duplicated imports + one unused
  repository import.

Remaining Sonar code smells (field injection, system clock in tests, unused
imports in other controllers) are non-blocking maintainability items left for
Feten — they do not affect the Quality Gate.

+++ microservices/scoring-service/src/main/java/tn/epos/scoring_service/service/EvaluateurDashboardService.java
@@ -383,7 +383,7 @@ public class EvaluateurDashboardService {
         }
 
         LocalDateTime debut     = rotation.getDebutCreneau();
-        LocalDateTime finReelle = debut.plusMinutes(DUREE_STATION_MIN + GRACE_PERIOD_MIN);
+        LocalDateTime finReelle = debut.plusMinutes((long) DUREE_STATION_MIN + GRACE_PERIOD_MIN);
 
         if (maintenant.isBefore(debut)) {
             return "A_VENIR";
@@ -544,7 +544,7 @@ public class EvaluateurDashboardService {
             return "TERMINE";
         }
         LocalDateTime debut     = rotation.getDebutCreneau();
-        LocalDateTime finReelle = debut.plusMinutes(DUREE_STATION_MIN + GRACE_PERIOD_MIN);
+        LocalDateTime finReelle = debut.plusMinutes((long) DUREE_STATION_MIN + GRACE_PERIOD_MIN);
 
         if (maintenant.isBefore(debut)) {
             return "A_VENIR";
